More tests
The polygon I have is actually a pathological case, so it may not make sense.  A lot of center points of cells come on the edge, so keys are not unique.  I'm thinking of adding a minimum jitter so that keys in the priority queue remain more or less unique.